pyproject.toml
[tool.poetry]
name = "tech-challenge-lgpd"
version = "1.0.0"
description = "Microserviço para gerenciar conformidade com a Lei Geral de Proteção de Dados (LGPD)"
authors = ["PosTech 5SOAT GRUPO 25 <postech-5soat-grupo-25@googlegroups.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.114.0"
httpx = "^0.27.2"
uvicorn = "^0.30.6"
pydantic = "^2.9.0"
databases = "^0.9.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"